Split EnemyBehaviour into itself, which contains the methods, and EnemyBehaviourOutputData, which contains the output data. Also created an interface for EnemyBehaviour, though I'm not sure if it's necessary. Moved the EnemyBehaviour classes into their own folder. Added a note to MovementInteractor
