+++ src/use_case/EnemyBehaviour/EnemyBehaviourInterface.java
@@ -1,0 +1,8 @@
+package use_case.EnemyBehaviour;
+
+import java.io.IOException;
+
+public interface EnemyBehaviourInterface {
+
+    EnemyBehaviourOutputData performRandomAction() throws IOException;
+}
